+++ include/linux/skmsg.h
@@ -595,7 +595,7 @@ static inline ssize_t sk_msg_first_len(struct sock *sk)
 
 static inline bool skb_bpf_strparser(const struct sk_buff *skb)
 {
-	unsigned long sk_redir = skb->_sk_redir;
+	uintptr_t sk_redir = skb->_sk_redir;
 
 	return sk_redir & BPF_F_STRPARSER;
 }
@@ -607,7 +607,7 @@ static inline void skb_bpf_set_strparser(struct sk_buff *skb)
 
 static inline bool skb_bpf_ingress(const struct sk_buff *skb)
 {
-	unsigned long sk_redir = skb->_sk_redir;
+	uintptr_t sk_redir = skb->_sk_redir;
 
 	return sk_redir & BPF_F_INGRESS;
 }
@@ -627,7 +627,7 @@ static inline void skb_bpf_set_redir(struct sk_buff *skb, struct sock *sk_redir,
 
 static inline struct sock *skb_bpf_redirect_fetch(const struct sk_buff *skb)
 {
-	unsigned long sk_redir = skb->_sk_redir;
+	uintptr_t sk_redir = skb->_sk_redir;
 
 	return (struct sock *)(sk_redir & BPF_F_PTR_MASK);
 }
