+++ net/ieee802154/6lowpan/reassembly.c
@@ -143,7 +143,7 @@ static int lowpan_frag_queue(struct lowpan_frag_queue *fq,
 	if (fq->q.flags == (INET_FRAG_FIRST_IN | INET_FRAG_LAST_IN) &&
 	    fq->q.meat == fq->q.len) {
 		int res;
-		unsigned long orefdst = skb->_skb_refdst;
+		uintptr_t orefdst = skb->_skb_refdst;
 
 		skb->_skb_refdst = 0UL;
 		res = lowpan_frag_reasm(fq, skb, prev_tail, ldev, refs);
