+++ net/ipv6/reassembly.c
@@ -218,7 +218,7 @@ static int ip6_frag_queue(struct net *net,
 
 	if (fq->q.flags == (INET_FRAG_FIRST_IN | INET_FRAG_LAST_IN) &&
 	    fq->q.meat == fq->q.len) {
-		unsigned long orefdst = skb->_skb_refdst;
+		uintptr_t orefdst = skb->_skb_refdst;
 
 		skb->_skb_refdst = 0UL;
 		err = ip6_frag_reasm(fq, skb, prev_tail, dev, refs);
