+++ net/openvswitch/actions.c
@@ -781,7 +781,7 @@ static void ovs_fragment(struct net *net, struct vport *vport,
 
 	if (key->eth.type == htons(ETH_P_IP)) {
 		struct rtable ovs_rt = { 0 };
-		unsigned long orig_dst;
+		uintptr_t orig_dst;
 
 		prepare_frag(vport, skb, orig_network_offset,
 			     ovs_key_mac_proto(key));
@@ -796,7 +796,7 @@ static void ovs_fragment(struct net *net, struct vport *vport,
 		ip_do_fragment(net, skb->sk, skb, ovs_vport_output);
 		refdst_drop(orig_dst);
 	} else if (key->eth.type == htons(ETH_P_IPV6)) {
-		unsigned long orig_dst;
+		uintptr_t orig_dst;
 		struct rt6_info ovs_rt;
 
 		prepare_frag(vport, skb, orig_network_offset,
